requirements.txt
torch==1.0.0
sentencepiece==0.1.83
numpy==1.17.1
tensorboardX==1.8
nltk==3.4.5
gdown
GitPython
pyrouge==0.1.3
rouge==0.3.2
